This commit fixes a very specific problem:
The #DEFINE DNNL_API appears throughout the OneDNN documentation in
function declarations between the type and function name. This
causes the cpp domain parser (a Sphinx built-in) to fail. Before
this I had to write a script to remove the "DNNL_API" string from
the XML source files to fix it.

This fix provides an interface for defining #DEFINEs that need to
be changed or removed:

``` python
breathe_pound_defines = {"DNNL_API",""}
```

This fix will cause it to be removed during the initial parsing phase
of the Doxygen XML. It only occurs in LinkedTextType fields, which is
where they were problematic.

In order to make this work, I had to access Sphinx config values from
the LinkedTextType class. The simplest way to do so was to create a
common sphinx app object that could be imported. Wiring in all of the
necessary inheritance would have been a massive undertaking.

Signed-off-by: Putnam, Kevin <kevin.putnam@intel.com>